Mouse glitch fix + added floor

+++ Real-time procedural destruction/Real-time procedural destruction/GameEngine/Input.h
@@ -21,7 +21,7 @@ namespace GameEngine
 		bool isKeyPressed(int _key); // Has a key been pressed this frame
 		bool isKeyReleased(int _key); // Has a key been released this frame
 
-		glm::vec2 getMousePos() { return m_mousePos; } // Getter for mouse cooridnate on the screen
+		glm::ivec2 getMousePos() { return m_mousePos; } // Getter for mouse cooridnate on the screen
 		glm::vec2 getMouseDelta() { return m_mouseDelta; } // Getter for the mouse velocity as it moves around
 		bool isMouseDown() { return m_mouseDown; } // If the mouse button is down
 		bool isMouseClicked() { return m_mouseDownClick; } // If the mouse button was clicked this frame
@@ -40,7 +40,7 @@ namespace GameEngine
 		std::vector<int> m_releasedKeys;
 
 		// Mouse Input
-		glm::vec2 m_mousePos;
+		glm::ivec2 m_mousePos;
 		glm::vec2 m_mouseDelta;
 		bool m_mouseDown;
 		bool m_mouseDownClick;
